refactor: rename files to kebab
