+++ src/adapters/cdn-bucket.ts
